Start distribution query tests
