+++ x/wasm/keeper/staking_test.go
@@ -460,9 +460,9 @@ func TestQueryStakingInfo(t *testing.T) {
 	require.NoError(t, err)
 	// first we pull out the data from chain response, before parsing the original response
 	var reflectRes testdata.ChainResponse
-	mustParse(t, res, &reflectRes)
+	mustUnmarshal(t, res, &reflectRes)
 	var bondedRes wasmvmtypes.BondedDenomResponse
-	mustParse(t, reflectRes.Data, &bondedRes)
+	mustUnmarshal(t, reflectRes.Data, &bondedRes)
 	assert.Equal(t, "stake", bondedRes.Denom)
 
 	// now, let's reflect a smart query into the x/wasm handlers and see if we get the same result
@@ -473,9 +473,9 @@ func TestQueryStakingInfo(t *testing.T) {
 	res, err = keeper.QuerySmart(ctx, maskAddr, reflectAllValidatorsBin)
 	require.NoError(t, err)
 	// first we pull out the data from chain response, before parsing the original response
-	mustParse(t, res, &reflectRes)
+	mustUnmarshal(t, res, &reflectRes)
 	var allValidatorsRes wasmvmtypes.AllValidatorsResponse
-	mustParse(t, reflectRes.Data, &allValidatorsRes)
+	mustUnmarshal(t, reflectRes.Data, &allValidatorsRes)
 	require.Len(t, allValidatorsRes.Validators, 1, string(res))
 	valInfo := allValidatorsRes.Validators[0]
 	// Note: this ValAddress not AccAddress, may change with #264
@@ -494,9 +494,9 @@ func TestQueryStakingInfo(t *testing.T) {
 	res, err = keeper.QuerySmart(ctx, maskAddr, reflectValidatorBin)
 	require.NoError(t, err)
 	// first we pull out the data from chain response, before parsing the original response
-	mustParse(t, res, &reflectRes)
+	mustUnmarshal(t, res, &reflectRes)
 	var validatorRes wasmvmtypes.ValidatorResponse
-	mustParse(t, reflectRes.Data, &validatorRes)
+	mustUnmarshal(t, reflectRes.Data, &validatorRes)
 	require.NotNil(t, validatorRes.Validator)
 	valInfo = *validatorRes.Validator
 	// Note: this ValAddress not AccAddress, may change with #264
@@ -516,9 +516,9 @@ func TestQueryStakingInfo(t *testing.T) {
 	res, err = keeper.QuerySmart(ctx, maskAddr, reflectNoValidatorBin)
 	require.NoError(t, err)
 	// first we pull out the data from chain response, before parsing the original response
-	mustParse(t, res, &reflectRes)
+	mustUnmarshal(t, res, &reflectRes)
 	var noValidatorRes wasmvmtypes.ValidatorResponse
-	mustParse(t, reflectRes.Data, &noValidatorRes)
+	mustUnmarshal(t, reflectRes.Data, &noValidatorRes)
 	require.Nil(t, noValidatorRes.Validator)
 
 	// test to get all my delegations
@@ -531,9 +531,9 @@ func TestQueryStakingInfo(t *testing.T) {
 	res, err = keeper.QuerySmart(ctx, maskAddr, reflectAllDelegationsBin)
 	require.NoError(t, err)
 	// first we pull out the data from chain response, before parsing the original response
-	mustParse(t, res, &reflectRes)
+	mustUnmarshal(t, res, &reflectRes)
 	var allDelegationsRes wasmvmtypes.AllDelegationsResponse
-	mustParse(t, reflectRes.Data, &allDelegationsRes)
+	mustUnmarshal(t, reflectRes.Data, &allDelegationsRes)
 	require.Len(t, allDelegationsRes.Delegations, 1)
 	delInfo := allDelegationsRes.Delegations[0]
 	// Note: this ValAddress not AccAddress, may change with #264
@@ -555,10 +555,10 @@ func TestQueryStakingInfo(t *testing.T) {
 	res, err = keeper.QuerySmart(ctx, maskAddr, reflectDelegationBin)
 	require.NoError(t, err)
 	// first we pull out the data from chain response, before parsing the original response
-	mustParse(t, res, &reflectRes)
+	mustUnmarshal(t, res, &reflectRes)
 
 	var delegationRes wasmvmtypes.DelegationResponse
-	mustParse(t, reflectRes.Data, &delegationRes)
+	mustUnmarshal(t, reflectRes.Data, &delegationRes)
 	assert.NotEmpty(t, delegationRes.Delegation)
 	delInfo2 := delegationRes.Delegation
 	// Note: this ValAddress not AccAddress, may change with #264
@@ -624,7 +624,7 @@ func TestQueryStakingPlugin(t *testing.T) {
 	raw, err := StakingQuerier(stakingKeeper, distributionkeeper.NewQuerier(distKeeper))(ctx, &query)
 	require.NoError(t, err)
 	var res wasmvmtypes.DelegationResponse
-	mustParse(t, raw, &res)
+	mustUnmarshal(t, raw, &res)
 	assert.NotEmpty(t, res.Delegation)
 	delInfo := res.Delegation
 	// Note: this ValAddress not AccAddress, may change with #264
